Create the abstract superclass for the element class and the omitted-element class.
