+++ packages/@markuplint/ml-core/src/ml-dom/tokens/element.ts
@@ -1,51 +1,40 @@
-import { AnonymousNode, Document } from '../';
-import { ContentModel, getNS } from '@markuplint/ml-spec';
-import { MLASTElement, MLToken } from '@markuplint/ml-ast';
-import { MLDOMAttribute, MLDOMElementCloseTag, MLDOMNode, MLDOMOmittedElement, MLDOMText, MLDOMToken } from './';
+import type { MLASTElement, MLToken } from '@markuplint/ml-ast';
+import type { MLDOMElementCloseTag, MLDOMText } from './';
 import { createNode, createSelector } from '../helper';
-import { IMLDOMElement } from '../types';
+import type { Document } from '../';
+import type { IMLDOMElement } from '../types';
+import MLDOMAbstractElement from './abstract-element';
+import MLDOMAttribute from './attribute';
 import MLDOMPreprocessorSpecificAttribute from './preprocessor-specific-attribute';
-import { RuleConfigValue } from '@markuplint/ml-config';
+import MLDOMToken from './token';
+import type { RuleConfigValue } from '@markuplint/ml-config';
 import { stringSplice } from '../../utils/string-splice';
 import { syncWalk } from '../helper/walkers';
 
 export default class MLDOMElement<T extends RuleConfigValue, O = null>
-	extends MLDOMNode<T, O, MLASTElement>
+	extends MLDOMAbstractElement<T, O, MLASTElement>
 	implements IMLDOMElement
 {
 	readonly type = 'Element';
-	readonly nodeName: string;
-	readonly attributes: (MLDOMAttribute | MLDOMPreprocessorSpecificAttribute)[];
 	readonly hasSpreadAttr: boolean;
-	readonly namespaceURI: string;
-	readonly isForeignElement: boolean;
 	readonly closeTag: MLDOMElementCloseTag<T, O> | null;
 	readonly selfClosingSolidus: MLDOMToken<MLToken> | null;
 	readonly endSpace: MLDOMToken<MLToken> | null;
-	readonly ownModels: Set<ContentModel> = new Set();
-	readonly childModels: Set<ContentModel> = new Set();
-	readonly descendantModels: Set<ContentModel> = new Set();
-	readonly isCustomElement: boolean;
 
 	readonly #tagOpenChar: string;
 	readonly #tagCloseChar: string;
 
-	#fixedNodeName: string;
-
 	constructor(astNode: MLASTElement, document: Document<T, O>) {
 		super(astNode, document);
-		this.nodeName = astNode.nodeName;
-		this.#fixedNodeName = astNode.nodeName;
-		this.attributes = astNode.attributes.map(attr =>
-			attr.type === 'html-attr' ? new MLDOMAttribute(attr) : new MLDOMPreprocessorSpecificAttribute(attr),
+		this.attributes.push(
+			...astNode.attributes.map(attr =>
+				attr.type === 'html-attr' ? new MLDOMAttribute(attr) : new MLDOMPreprocessorSpecificAttribute(attr),
+			),
 		);
 		this.hasSpreadAttr = astNode.hasSpreadAttr;
 		this.selfClosingSolidus = astNode.selfClosingSolidus ? new MLDOMToken(astNode.selfClosingSolidus) : null;
 		this.endSpace = astNode.endSpace ? new MLDOMToken(astNode.endSpace) : null;
-		this.namespaceURI = astNode.namespace;
-		this.isForeignElement = this.namespaceURI !== 'http://www.w3.org/1999/xhtml';
 		this.closeTag = astNode.pearNode ? createNode(astNode.pearNode, document, this) : null;
-		this.isCustomElement = astNode.isCustomElement;
 
 		this.#tagOpenChar = astNode.tagOpenChar;
 		this.#tagCloseChar = astNode.tagCloseChar;
@@ -54,9 +43,9 @@ export default class MLDOMElement<T extends RuleConfigValue, O = null>
 	get raw() {
 		let fixed = this.originRaw;
 		let gap = 0;
-		if (this.nodeName !== this.#fixedNodeName) {
-			fixed = stringSplice(fixed, this.#tagOpenChar.length, this.nodeName.length, this.#fixedNodeName);
-			gap = gap + this.#fixedNodeName.length - this.nodeName.length;
+		if (this.nodeName !== this.fixedNodeName) {
+			fixed = stringSplice(fixed, this.#tagOpenChar.length, this.nodeName.length, this.fixedNodeName);
+			gap = gap + this.fixedNodeName.length - this.nodeName.length;
 		}
 		for (const attr of this.attributes) {
 			const startOffset =
@@ -72,33 +61,11 @@ export default class MLDOMElement<T extends RuleConfigValue, O = null>
 		return fixed;
 	}
 
-	get childNodes(): AnonymousNode<T, O>[] {
-		const astChildren = this._astToken.childNodes || [];
-		return astChildren.map(node => this.nodeStore.getNode<typeof node, T, O>(node));
-	}
-
-	get ns() {
-		if (/[a-z]:[a-z]/i.test(this.nodeName)) {
-			return this.nodeName.split(':')[0];
-		}
-		return getNS(this.namespaceURI);
-	}
-
-	get nameWithNS() {
-		if (/[a-z]:[a-z]/i.test(this.nodeName)) {
-			return this.nodeName;
-		}
-		const ns = getNS(this.namespaceURI);
-		if (ns === 'html') {
-			return this.nodeName.toLowerCase();
-		}
-		return `${ns}:${this.nodeName}`;
-	}
-
 	querySelectorAll(selector: string) {
 		const matchedNodes: (MLDOMElement<T, O> | MLDOMText<T, O>)[] = [];
+		const selecor = createSelector(selector);
 		syncWalk(this.childNodes, node => {
-			if (node.type === 'Element' && node.matches(selector)) {
+			if (node.type === 'Element' && selecor.match(node, this)) {
 				matchedNodes.push(node);
 			}
 			if (selector === '#text' && node.type === 'Text' && !node.isWhitespace()) {
@@ -108,120 +75,11 @@ export default class MLDOMElement<T extends RuleConfigValue, O = null>
 		return matchedNodes;
 	}
 
-	closest(selector: string) {
-		let el: MLDOMElement<T, O> | MLDOMOmittedElement<T, O> | null = this;
-
-		do {
-			if (el.matches(selector)) {
-				return el;
-			}
-			el = el.parentNode;
-		} while (el !== null && el.type === 'Element');
-		return null;
-	}
-
-	getAttributeToken(attrName: string) {
-		const attrs: (MLDOMAttribute | MLDOMPreprocessorSpecificAttribute)[] = [];
-		attrName = attrName.toLowerCase();
-		for (const attr of this.attributes) {
-			if (attr.potentialName === attrName) {
-				attrs.push(attr);
-			}
-		}
-		return attrs;
-	}
-
-	getAttribute(attrName: string) {
-		attrName = attrName.toLowerCase();
-		for (const attr of this.attributes) {
-			if (attr.potentialName === attrName) {
-				if (attr.attrType === 'html-attr') {
-					return attr.value ? attr.value.raw : null;
-				} else {
-					return attr.potentialValue;
-				}
-			}
-		}
-		return null;
-	}
-
-	hasAttribute(attrName: string) {
-		return !!this.getAttributeToken(attrName).length;
-	}
-
-	matches(selector: string): boolean {
-		// @ts-ignore TODO: To improvement testable object
-		return createSelector(selector).match(this);
-	}
-
-	fixNodeName(name: string) {
-		this.#fixedNodeName = name;
-	}
-
-	getChildElementsAndTextNodeWithoutWhitespaces() {
-		const filteredNodes: (MLDOMElement<T, O> | MLDOMText<T, O>)[] = [];
-		this.childNodes.forEach(node => {
-			if (node.type === 'Element') {
-				filteredNodes.push(node);
-			}
-			if (node.type === 'Text' && !node.isWhitespace()) {
-				filteredNodes.push(node);
-			}
-			if (node.type === 'OmittedElement') {
-				const children = node.getChildElementsAndTextNodeWithoutWhitespaces();
-				filteredNodes.push(...children);
-			}
-		});
-		return filteredNodes;
-	}
-
-	/**
-	 * This element has "Preprocessor Specific Block". In other words, Its children are potentially mutable.
-	 */
-	hasMutableChildren() {
-		return this.childNodes.some(node => node.type === 'PSBlock');
-	}
-
-	isDescendantByUUIDList(uuidList: string[]) {
-		let el: MLDOMElement<T, O> | MLDOMOmittedElement<T, O> | null = this.parentNode;
-
-		if (el === null) {
-			return false;
-		}
-
-		do {
-			if (uuidList.includes(el.uuid)) {
-				return true;
-			}
-			el = el.parentNode;
-		} while (el !== null && el.type === 'Element');
-		return false;
-	}
-
 	getNameLocation() {
 		return {
 			offset: this.startOffset,
 			line: this.startLine,
 			col: this.startCol + this.#tagOpenChar.length,
 		};
-	}
-
-	get classList() {
-		const classList: string[] = [];
-		const classAttrs = this.getAttributeToken('class');
-		for (const classAttr of classAttrs) {
-			const value = classAttr.attrType === 'html-attr' ? classAttr.value.raw : classAttr.potentialValue;
-			classList.push(
-				...value
-					.split(/\s+/g)
-					.map(c => c.trim())
-					.filter(c => c),
-			);
-		}
-		return classList;
-	}
-
-	get id() {
-		return this.getAttribute('id') || '';
 	}
 }
